refactor delete function in service due the "not-null property references a null or transient value" error

+++ common/src/main/java/com/cqlybest/common/service/AccountService.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.cqlybest.common.bean.Account;
 import com.cqlybest.common.dao.AccountDao;
@@ -19,17 +20,20 @@ public class AccountService {
    * 
    * @param account
    */
+  @Transactional
   public void add(Account account) {
     accountDao.saveOrUpdate(account);
   }
 
   /**
    * 删除账号
-   * 
-   * @param ccount
    */
-  public void delete(Account account) {
-    accountDao.delete(account);
+  @Transactional
+  public void delete(Integer id) {
+    Account account = accountDao.findById(id);
+    if (account != null) {
+      accountDao.delete(account);
+    }
   }
 
   /**
@@ -55,6 +59,7 @@ public class AccountService {
    * 
    * @param account
    */
+  @Transactional
   public void modify(Account account) {
     accountDao.saveOrUpdate(account);
   }
